requirements.txt
click==8.0.1
humanize==3.9.0
keyboard==0.13.5
PyGithub==1.55
python-decouple==3.4
python-gitlab==2.8.0
rich==10.4.0
idna==2.10
